Moved orionldAttributeTreat to its own file
